support heading IDs
If an ATX heading is followed by {#id}, then its HTML
form will have an "id" attribute. For example,

    # Names {#names}

will render as

    <h1 id="names">Names</h1>

To keep CommonMark compatibility the default, I added an option
to the parser, which is now exported. I modeled the API on
go/doc/comments.

To test, I extended the testing framework to allow

    key: value

lines in the txtar comment section, which are interpreted as
parser options.