render markdown
Add ToMarkdown, which renders a Document as Markdown.

It is essentially a Markdown pretty-printer, although most of the time
it tries to be faithful to the input.

This doesn't yet handle all the strange cases in the 0.30 spec tests,
but it does work on the copies of the Go 1.20 and 1.21 release notes
included here, producing identical output.