markdown: fix/break table escaping
In any sane language, if \\ means \ and | means |
and \\ | means \ |, then \\| (without the space) would mean \|.
But this is not a sane language.
This is GitHub Flavored Markdown, and since it is Markdown,
that means any time there is a chance to make a mistake and
do something unprincipled, you do it.

In this case, \| and \\| both mean |, \\\| and \\\\| both mean \|, and so on.

Fixes #9.